refactor(assistant): split Ollama provider handler test
- move the LLMock Ollama-compatible provider coverage into a provider-local test
- keep handler.test focused on request streaming, UI messages, and Workers AI gateway behavior
- reduce handler.test hotspot score from 30.4 to 26.1 and cyclomatic total from 21 to 12
